Linux 2.6 Open Source Gold Release
Added support for Reproducible Enclave Build using Docker file.
Added support for Intel AVX-512 instructions and Intel SHA Extensions New Instructions (SHA-NI) in trusted libraries.
Support both EPID and ECDSA based quote for quoting related interfaces in sgx_uae_service library.
Updated key exchange library to support both EPID and ECDSA based remote attestation.
Support new interface to check platform information blob from remote attestation response message.
Fixed bugs.

+++ common/inc/internal/se_map.h
@@ -55,7 +55,7 @@ extern "C" {
 #endif
 
 /* Map a file to memory and get its size. */
-map_handle_t* map_file(se_file_handle_t fd, uint32_t *size);
+map_handle_t* map_file(se_file_handle_t fd, off_t *size);
 
 /* Unmap a mapped file. */
 void unmap_file(map_handle_t* mh);
